refactor: move VmDetail's VmConsoles to separate file
